Refactor DividedHasher to BodySuffixHasher
