--- circuits-circom/circuits/common-v2/divided_hasher_10752.circom
@@ -1,6 +1,0 @@
-pragma circom 2.1.5;
-
-include "./helpers/divided_hasher.circom";
-
-// TOTAL CONSTRAINTS: 5316745
-component main = DividedHasher(10752);
