+++ circuits-circom/circuits/common-v2/test/mocks/test_body_suffix_hasher.circom
@@ -1,0 +1,3 @@
+include "../../body_suffix_hasher.circom";
+
+component main = BodySuffixHasher(320);
