--- circuits-circom/circuits/common-v2/test/mocks/test_divided_hasher.circom
@@ -1,3 +1,0 @@
-include "../../helpers/divided_hasher.circom";
-
-component main = DividedHasher(320);
