+++ circuits-circom/circuits/garanti/garanti_body_suffix_hasher.circom
@@ -1,0 +1,6 @@
+pragma circom 2.1.5;
+
+include "../common-v2/body_suffix_hasher.circom";
+
+// TOTAL CONSTRAINTS: 5316745
+component main = BodySuffixHasher(10752);
